+++ nengo_loihi/simulator.py
@@ -57,11 +57,12 @@ class Simulator:
         want to build the network manually, or you want to inject build
         artifacts in the model before building the network, then you can
         pass in a `.Model` instance.
-    precompute : bool, optional (Default: False)
+    precompute : bool, optional (Default: None)
         Whether model inputs should be precomputed to speed up simulation.
         When *precompute* is False, the simulator will be run one step
         at a time in order to use model outputs as inputs in other parts
-        of the model.
+        of the model. By default, the simulator will choose ``True`` if it works for
+        your model, and ``False`` otherwise.
     target : str, optional (Default: None)
         Whether the simulator should target the emulator (``'sim'``) or
         Loihi hardware (``'loihi'``). If None, *target* will default to
@@ -96,7 +97,7 @@ class Simulator:
         dt=0.001,
         seed=None,
         model=None,
-        precompute=False,
+        precompute=None,
         target=None,
         progress_bar=None,
         remove_passthrough=True,
@@ -104,7 +105,6 @@ class Simulator:
     ):
         # initialize values used in __del__ and close() first
         self.closed = True
-        self.precompute = precompute
         self.network = network
         self.sims = OrderedDict()
         self._run_steps = None
@@ -138,7 +138,9 @@ class Simulator:
 
         # determine how to split the host into one, two or three models
         self.model.split = Split(
-            network, precompute=precompute, remove_passthrough=remove_passthrough
+            network,
+            precompute=False if precompute is None else precompute,
+            remove_passthrough=remove_passthrough,
         )
 
         # Build the network into the model
@@ -153,8 +155,24 @@ class Simulator:
             self.model.seeded[conn] = False
             self.model.build(conn)
 
-        if len(self.model.host_pre.params):
-            assert precompute
+        # Create host_pre and host simulators if necessary
+        self.precompute = precompute
+        if precompute is None:
+            # If there is no host then all objects must be on the chip,
+            # which is precomputable in the sense that
+            # no communication has to happen with the host.
+            self.precompute = len(self.model.host.params) == 0
+        elif len(self.model.host_pre.params) == 0 and precompute:
+            warnings.warn(
+                "No precomputable objects. Setting precompute=True has no effect."
+            )
+        elif len(self.model.host.params) == 0 and not precompute:
+            warnings.warn(
+                "Model is pre-computable. Setting precompute=False may slow execution."
+            )
+
+        if len(self.model.host_pre.params) > 0:
+            assert self.precompute
             self.sims["host_pre"] = nengo.Simulator(
                 network=None,
                 dt=self.dt,
@@ -162,12 +180,8 @@ class Simulator:
                 progress_bar=False,
                 optimize=False,
             )
-        elif precompute:
-            warnings.warn(
-                "No precomputable objects. Setting " "precompute=True has no effect."
-            )
 
-        if len(self.model.host.params):
+        if len(self.model.host.params) > 0:
             self.sims["host"] = nengo.Simulator(
                 network=None,
                 dt=self.dt,
@@ -175,13 +189,6 @@ class Simulator:
                 progress_bar=False,
                 optimize=False,
             )
-        elif not precompute:
-            # If there is no host and precompute=False, then all objects
-            # must be on the chip, which is precomputable in the sense that
-            # no communication has to happen with the host.
-            # We could warn about this, but we want to avoid people having
-            # to specify `precompute` unless they absolutely have to.
-            self.precompute = True
 
         self._probe_outputs = self.model.params
         self.data = SimulationData(self._probe_outputs)